Atribuição da Data diretamente no construtor de Author e remoção do controller
