+++ src/br/com/zup/nossacasacodigo/test/AuthorServiceTest.java
@@ -4,7 +4,7 @@ import org.junit.Assert;
 import org.junit.Test;
 
 import br.com.zup.nossacasacodigo.author.Author;
-import br.com.zup.nossacasacodigo.controllers.AuthorService;
+import br.com.zup.nossacasacodigo.database.CollectionOfAuthors;
 import junit.framework.TestCase;
 
 public class AuthorServiceTest extends TestCase {
@@ -15,15 +15,16 @@ public class AuthorServiceTest extends TestCase {
 		//texto de exemplo com exatamente 400 caracteres
 		String description = "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Proin et placerat purus. Orci varius natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus. Pellentesque vulputate libero tortor. Fusce arcu felis, rhoncus eget est in, suscipit venenatis orci. Vestibulum eget feugiat justo. Morbi sit amet felis finibus, imperdiet metus eget, pellentesque risus. Nulla at porttitor turpis.";
 		
-		AuthorService service = new AuthorService();
+		CollectionOfAuthors database = new CollectionOfAuthors();
 		
-		Author newAuthor = service.newAuthor(name, email, description);
-		
+		Author newAuthor = new Author(name, email, description);
+		database.addNewAuthor(newAuthor);
+				
 		Assert.assertNotNull(newAuthor);
 		Assert.assertEquals( newAuthor.getName(), name);
 		Assert.assertEquals(newAuthor.getEmail(), email);
 		Assert.assertEquals(newAuthor.getDescription(), description);
-		Assert.assertEquals(1, service.quantityOfAuthors());
+		Assert.assertEquals(1, quantityOfAuthors(database));
 	}
 	
 	@Test
@@ -38,11 +39,13 @@ public class AuthorServiceTest extends TestCase {
 		//texto de exemplo com menos de 400 caracteres
 		String descriptionTwo = "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Orci varius natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus. Pellentesque vulputate libero tortor. Fusce arcu felis, rhoncus eget est in, suscipit venenatis orci.  Nulla at porttitor turpis.";
 		
-		AuthorService service = new AuthorService();
+		CollectionOfAuthors database = new CollectionOfAuthors();
 		
-		Author newAuthorOne = service.newAuthor(name, email, description);
-		Author newAuthorTwo = service.newAuthor(nameTwo, emailTwo, descriptionTwo);
+		Author newAuthorOne = new Author(name, email, description);
+		Author newAuthorTwo = new Author(nameTwo, emailTwo, descriptionTwo);
 		
+		database.addNewAuthor(newAuthorOne);
+		database.addNewAuthor(newAuthorTwo);
 		Assert.assertNotNull(newAuthorOne);
 		Assert.assertEquals( newAuthorOne.getName(), name);
 		Assert.assertEquals(newAuthorOne.getEmail(), email);
@@ -53,7 +56,16 @@ public class AuthorServiceTest extends TestCase {
 		Assert.assertEquals(newAuthorTwo.getEmail(), emailTwo);
 		Assert.assertEquals(newAuthorTwo.getDescription(), descriptionTwo);
 		
-		Assert.assertEquals(2, service.quantityOfAuthors());
+		
+		Assert.assertEquals(2, quantityOfAuthors(database));
+	}
+	
+	private int quantityOfAuthors(CollectionOfAuthors database) {
+		int quantity = 0;
+		for (Author print : database.getCollection().values()) { 
+			quantity++;
+		}
+		return quantity;
 	}
 	
 }
